Move shape template generation into shapes calculator
